begin to use xmltree to parse graphml
It seems like we need to read the graphml file into json, and that there does not seem to be a good
way to do that using the "xml_string_to_json" function because that function seems to parse the
graphml into a json structure that I don't recognize.
It may make sense to use rustworkx, which would (hopefully) mirror the python implementation which
uses networkx.