`rpc`: remove kubectl
Had to do an annoying dance because it seems like stream() wants to convert any json it sees into
python primitives. There's gotta be a better way to deal with that, but I came up with crudely
checking if we are getting a list or dict and literal_eval-ing it. Maybe I'm missing something.
Otherwise, seems to work.