constants: make the default namespace "default"
The general agreement we came to was to use "default" as the default namespace. I took a moment to
search the code base for instances of hard-coded "warnet" namespaces, and I could only really find
one. It was in the commander file -- specifically in the network flag area. I updated this in another commit.