k8s: add `continue` to `wait_for_init`
When this function encounters a matching pod, it attempts to retrieve that pod's
init_container_statuses. The problem seems to be that this metadata is not available (perhaps
because the pod is not read?). This means that the metadata is set to None which is not iterable and
therefore causes a crash.

This fix allows the program to proceed to the next element in the for loop stream if the init_container_statuses metadata
is not ready (aka is None).